Constructors, Destructors: Rule of All or Nothing
Right on the first class template `List`, we see a default constructor
that is explicitly implemented to do nothing, i.e. we should use the
keyword `default`. The same goes for the destructor. Since that one is
virtual, we can not leave it away. That means we should also have a look
at the move and copy operations.

`List` contains only a `vector`, which is fully copy/moveable, so we can
default all special members there. `Statement` is empty, so it is
obvious what the default does and it is sensible to loosen the reigns of
the rule a bit and only default the virtual destructor. For all other
classes except `CompoundStatement` the Rule of Zero applies, they need
not be changed.

`CompoundStatement` itself has a nontrivial destructor due to the fact
that it manages the lifetimes of the `List` elements. If we look closer
it becomes apparent that were we to copy a `CompoundStatement` with a
nonempty `List`, the pointers in that list would get copied as well and
deleted twice eventually.

The move constructor will work, but not the move assignment since the
old contents will not be deleted and therefore leak. So default and move
constructor as well as the destructor can be defaulted, the rest has to
be deleted.