Single line blocks
Blocks that consist of a single line, e.g. of function bodies and for
loops, should be wrapped in their own curly braces and put on their own
line. Putting things on their own line visibly separates the two
separate parts of the loop - the header and the loop body. Adding the
braces even on one-liners prevent errors that arise from adding more
lines to the apparent block without adding the braces then.

This is somewhat a matter of taste and coding style, but many style
guides stick at least to the own line for loop bodies. Most people seem
to favor the separation over terseness.