Remove superfluous inline declarations
In the past, the keyword `inline` has been a hint to the compiler that
it might try to inline a function. Modern compilers usually ignore it
completely and it is only used to obey the One Definition Rule. In other
words, use it only if you feel the need to define non-template functions
outside a class definition.

In this code, all the functions declared `inline` are defined inside a
class definition, which means that they are already implicitly declared
inline. Therefore the explicit `inline` is superfluous and we should
simply remove it.