Make data members private
The member variables of `Declaration` and all subclasses of `Statement`
are public. This does not seem to be necessary, and since the classes
are more than plain data containers their members should be private. In
fact I like to distinguish classes versus data structures by using the
keywords `class` and `struct` accordingly, but I will leave those as
they are in this case.

Another case is the `List` base of `CompoundStatement` which is in fact
more a data member than a base class, so I should make it private, too.
However, the `main()` function calls `Append`, so it's not that trivial.