Take ownership explicitly
Fix the interface of the `Append()` method by explicitly demanding that
any client pass it a `unique_ptr`. It will also make the implementation
of that method much more natural. Doing that will enable us to use
`make_unique` instead of `new` in the `main()` function - so in addition
to the clearer interface, we also get some exception safety for free.